backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0
python-jose[cryptography]==3.3.0
bcrypt>=4.1.2
python-multipart==0.0.6
sqlalchemy==2.0.25
alembic==1.13.1
cryptography>=42.0.0
python-dotenv>=1.0.0
osmnx>=1.7.2
geopandas>=0.14.0
folium>=0.14.0
requests>=2.31.0
httpx>=0.25.0
